organizing paths and files
