+++ app/api/agendamento.js
@@ -8,10 +8,11 @@ export default (req, res) => {
         const agendamento = JSON.parse(body)
         console.log(agendamento)
         res.writeHead(200, {
-            'Content-Type': 'application/json',
+            'Content-Type': 'application/json; charset=utf-8',
             'Access-Control-Allow-Origin': '*',
             'Access-Control-Allow-Methods': 'GET, POST',
-            'Access-Control-Allow-Headers': 'Content-Type'
+            'Access-Control-Allow-Headers': 'Content-Type',
+            'X-Content-Type-Options': 'nosniff'
         })
         res.end(JSON.stringify({ message: 'Agendamento recebido' }))
     })
