+++ app/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+<head>
+    <title>teste</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+</head>
+<body>
+    <input type="text" id="nome" placeholder="nome" title="nome"> 
+    <input type="number" id="celular" placeholder="numero" title="celular"> 
+    <button onclick="handleButtonClick()">agendar</button>
+    <script src="./agendar.js"></script>
+</body>
+</html>
