--- index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-</head>
-<body>
-    <input type="text" id="nome"> 
-    <input type="number" id="celular"> 
-    <input type="checkbox" id="pibid"> 
-    <button onclick="handleButtonClick()">agendar</button>
-    <script src="./agendar.js"></script>
-</body>
-</html>
