+++ package.json
@@ -22,12 +22,6 @@
     "url": "https://github.com/mdsreq-fga-unb/2023.2-AngelaCabeleireira/issues"
   },
   "homepage": "https://github.com/mdsreq-fga-unb/2023.2-AngelaCabeleireira#readme",
-  "dependencies": {
-    "puppeteer": "^13.7.0",
-    "qrcode-terminal": "^0.12.0",
-    "serve": "^14.2.1",
-    "whatsapp-web.js": "^1.23.0"
-  },
   "devDependencies": {
     "jest": "^29.7.0"
   }
